add guard to remove the archive dir

+++ tests/metrics/conftest.py
@@ -126,7 +126,8 @@ def local_collector(request):
                if hasattr(request.config.option, 'password') else \
                'Testing123@'
     archive_dir = os.path.join(os.environ['TEST_ARTIFACTS'], 'archives')
-    shutil.rmtree(archive_dir)
+    if os.path.isdir(archive_dir):
+        shutil.rmtree(archive_dir)
     os.mkdir(archive_dir)
     collector = LocalCollector(archive_dir)
     collector.set_deployment(deployment)
